Split up API v2
Images and documents endpoints are now defined in their respective apps
